+++ vulkan-rust/crates/gamedata/src/locode.rs
@@ -4,15 +4,13 @@ use crate::vector::Vec3;
 
 pub const ROOT_LOCATION_CODE: u64 = 0b1;
 
-pub struct LocationCode {
-
-}
+pub struct LocationCode {}
 
 impl LocationCode {
     pub fn from_vec(vector: Vec3) -> u64 {
         u64::from(vector.x.is_sign_positive()) << 2
-        | u64::from(vector.y.is_sign_positive()) << 1
-        | u64::from(vector.z.is_sign_positive())
+            | u64::from(vector.y.is_sign_positive()) << 1
+            | u64::from(vector.z.is_sign_positive())
     }
 
     pub fn child_index_to_vec(child_index: u64) -> Vec3 {
@@ -21,9 +19,9 @@ impl LocationCode {
         let code = child_index as u8;
 
         glm::vec3(
-            -0.5 + f64::from(code >> 2),
-            -0.5 + f64::from(code >> 1 & 0b1),
-            -0.5 + f64::from(code & 0b1),
+            -0.5 + f32::from(code >> 2),
+            -0.5 + f32::from(code >> 1 & 0b1),
+            -0.5 + f32::from(code & 0b1),
         )
     }
 }
